requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
alembic==1.13.1  # Added for database migrations
python-dateutil
httpx>=0.24.0,<1.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0